// sgo-Mobile/src/app/shared/interfaces/Chamado.ts
export interface ChamadoInterface {
    id: string;
    statusChamado: string;
    protocolo: string;
    ultimo_evento: string;
    categoriaProblema: any;
    usuario: any;
    feedbackUsuario: number;
    envolveAreaComum: boolean;
    areaComum: any;
    apartamento: any;
    descricao: string;
    novosEventos: boolean;
    img: string;
}